+++ packages/gestalt/src/index.js
@@ -22,8 +22,6 @@ import Letterbox from './Letterbox.js';
 import Link from './Link.js';
 import Mask from './Mask.js';
 import Masonry from './Masonry.js';
-import MasonryBeta from './MasonryBeta.js';
-import MasonryInfiniteBeta from './MasonryInfiniteBeta.js';
 import MasonryDefaultLayout from './layouts/MasonryLayout.js';
 import MasonryUniformRowLayout from './layouts/UniformRowLayout.js';
 import Modal from './Modal.js';
@@ -70,8 +68,6 @@ export {
   Link,
   Mask,
   Masonry,
-  MasonryBeta,
-  MasonryInfiniteBeta,
   MasonryDefaultLayout,
   MasonryUniformRowLayout,
   Modal,